# docs/tech/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WebSR API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models/download": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Download Status",
        "description": "下载作业状态：`{status, entry_id, received_bytes, total_bytes, model_id, conversion_triggered, error}`。\n\n`status=completed` 后转换可能仍在进行——转换进度看\n`GET /api/models/{model_id}/convert`（T-807 既有端点）。",
        "operationId": "download_status_api_models_download_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Download Status Api Models Download Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Start Download",
        "description": "从目录下载模型（**异步**）：下载 → 登记 → 自动进入应用内转换。",
        "operationId": "start_download_api_models_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Start Download Api Models Download Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_models_get",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelOut"
                  },
                  "title": "Response List Models Api Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/import": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Import Model",
        "operationId": "import_model_api_models_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_model_api_models_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/export": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Export Model",
        "operationId": "export_model_api_models__model_id__export_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/convert": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Convert Model",
        "description": "触发 `.pth` / `.safetensors` → `.onnx` 的离线转换（**异步**，立即返回作业号）。\n\n转换在独立环境里以子进程执行（ADR-003：应用内不含 torch），导出实测数十秒，\n同步等待会顶穿前端超时——故立即返回，用 `GET` 取状态。",
        "operationId": "convert_model_api_models__model_id__convert_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Convert Model Api Models  Model Id  Convert Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Convert Status",
        "description": "转换作业状态与结果（与触发同一路径，`GET` 取状态）。\n\n返回 `{status, job_id, model_id, started_at, finished_at, result, error, availability}`。\n`result.model_id` 是**转换产物**登记成的新模型 id（与源模型 id 不同）。",
        "operationId": "convert_status_api_models__model_id__convert_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Convert Status Api Models  Model Id  Convert Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Model",
        "operationId": "delete_model_api_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_model_api_models_import_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "scale": {
            "type": "integer",
            "title": "Scale"
          },
          "min_vram_mb": {
            "type": "integer",
            "title": "Min Vram Mb"
          },
          "backends": {
            "type": "string",
            "title": "Backends"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "companion_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companion File"
          }
        },
        "type": "object",
        "required": [
          "name",
          "format",
          "scale",
          "min_vram_mb",
          "backends",
          "file"
        ],
        "title": "Body_import_model_api_models_import_post"
      },
...
      "DownloadIn": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          }
        },
        "type": "object",
        "required": [
          "entry_id"
        ],
        "title": "DownloadIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelCapabilities": {
        "properties": {
          "supports_fp16": {
            "type": "boolean",
            "title": "Supports Fp16"
          },
          "supports_batch": {
            "type": "boolean",
            "title": "Supports Batch"
          },
          "supports_tile0": {
            "type": "boolean",
            "title": "Supports Tile0"
          },
          "has_tensorrt": {
            "type": "boolean",
            "title": "Has Tensorrt"
          },
          "is_generative": {
            "type": "boolean",
            "title": "Is Generative"
          },
          "num_inference_steps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Inference Steps"
          },
          "requires_prompt": {
            "type": "boolean",
            "title": "Requires Prompt"
          }
        },
        "type": "object",
        "required": [
          "supports_fp16",
          "supports_batch",
          "supports_tile0",
          "has_tensorrt",
          "is_generative",
          "num_inference_steps",
          "requires_prompt"
        ],
        "title": "ModelCapabilities"
      },
      "ModelConversion": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "available",
          "reason"
        ],
        "title": "ModelConversion",
        "description": "`.pth` / `.safetensors` 的离线转换可用性（T-807）。\n\n非这两种格式时 `ModelOut.conversion` 为 `null`——\"不需要转换\"与\n\"需要但环境没装\"是两回事，用一个可空对象表达而不是两个布尔。"
      },
      "ModelOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "params_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params Count"
          },
          "scale": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "min_vram_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Vram Mb"
          },
          "supported_backends": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supported Backends"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities"
          },
          "companion": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companion"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "conversion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelConversion"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "architecture",
          "description",
          "format",
          "path",
          "sha256",
          "size_bytes",
          "params_count",
          "scale",
          "license",
          "source",
          "min_vram_mb",
          "supported_backends",
          "capabilities",
          "companion",
          "available",
          "unavailable_reason",
          "status"
        ],
        "title": "ModelOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}